Resend export win from the details page
