+++ src/client/actions.js
@@ -314,8 +314,7 @@ export const COPY_TASK_DETAILS_LOADED = 'COPY_TASK_DETAILS_LOADED'
 
 export const INVESTMENT__PROJECT_LOADED = 'INVESTMENT__PROJECT_LOADED'
 
-export const EXPORT_WINS__CONFIRMED_LOADED = 'EXPORT_WINS__CONFIRMED_LOADED'
-export const EXPORT_WINS__DETAILS_LOADED = 'EXPORT_WINS__DETAILS_LOADED'
+export const EXPORT_WIN_RESENT__SUCCESS = 'EXPORT_WIN_RESENT__SUCCESS'
 
 export const PAGINATED_RESOURCE__GO_TO_PAGE = 'PAGINATED_RESOURCE__GO_TO_PAGE'
 export const PAGINATED_RESOURCE__ON_SUCCESS = 'PAGINATED_RESOURCE__ON_SUCCESS'
