+++ src/client/modules/ExportWins/Details/state.js
@@ -1,0 +1,3 @@
+export const TASK_RESEND_EXPORT_WIN = 'TASK_RESEND_EXPORT_WIN'
+export const ID = 'ResendExportWin'
+export const state2props = (state) => state[ID]
