+++ src/client/modules/ExportWins/Details/tasks.js
@@ -1,0 +1,4 @@
+import { apiProxyAxios } from '../../../components/Task/utils'
+
+export const resendExportWin = (id) =>
+  apiProxyAxios.post(`/v4/export-win/${id}/resend-win`)
