+++ src/client/reducers.js
@@ -181,6 +181,8 @@ import getInteractionReducer from './modules/Interactions/InteractionDetails/red
 import { PREVIEW_QUOTE_ID } from './modules/Omis/state'
 import orderQuoteReducer from './modules/Omis/reducer'
 
+import { ResendExportWin } from './modules/ExportWins/Details/ResendExportWin'
+
 export const reducers = {
   tasks,
   [FLASH_MESSAGE_ID]: flashMessageReducer,
@@ -202,6 +204,7 @@ export const reducers = {
   ...ContactForm.reducerSpread,
   ...Form.reducerSpread,
   ...FieldAddAnother.reducerSpread,
+  ...ResendExportWin.reducerSpread,
   [DNB_CHECK_ID]: dnbCheckReducer,
   [INVESTMENT_OPPORTUNITIES_LIST_ID]: investmentOpportunitiesListReducer,
   [INVESTMENT_OPPORTUNITIES_DETAILS_ID]: investmentOpportunitiesDetailsReducer,
